requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0
Authlib==1.3.0
httpx==0.24.1
Jinja2==3.1.3
python-multipart==0.0.9
itsdangerous==2.1.2
google-api-python-client==2.120.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0